+++ src/extension/fixes/optimizeUE4.ts
@@ -1,143 +1,58 @@
 /**
- * This fix optimizes the UE workspace by removing any project specific references and adding basic Intellisense
- * 
- * To add preincludes, without polluting the UE workspace, we need to set them in the '.code-workspace' file.
+ * This fix optimizes the UE workspace by setting the intellisense to Tag Parser.
+ * Tag parser Intellisense removes red squiggles but keeps navigation of the UE source.
  */
 
 
-import type {WorkspaceConfiguration} from "vscode";
+import * as vscode from "vscode"
 
 import { ProjectUE4 } from "../../project/projectUE4";
-import type { CCppConfigurationJson } from "../../project/ntypes";
-import * as consts from "../../consts";
-import { readStringFromFile } from "../../shared";
 
 import * as console from "../../console";
 
 
-export async function fixUE4Optimization(project: ProjectUE4) {
+const SETTING_TAG_PARSER = "Tag Parser";
+const SETTING_INTELLISENSE_ENGINE = "intelliSenseEngine";
+
+
+export async function fixUEOptimization(project: ProjectUE4, isOptimizationEnabled: boolean) {
     console.log("Attempting to fix UE workspace optimization.");
 
-    const ue4CCppPropertiesFirstConfig = project.getFirstCCppPropertiesConfiguration(project.ue4WorkspaceKey);
-    const allUE4CppPropertiesConfigs = project.getCCppConfigurationsFromWorkspace(project.ue4WorkspaceKey);
 
-    if (!ue4CCppPropertiesFirstConfig || !allUE4CppPropertiesConfigs) {
-        console.error("Couldn't get UE's c_cpp_properties.json configs.");
-        return;
+    const configScope = project.ue4WorkspaceFolder;
+    const cppConfig = vscode.workspace.getConfiguration("C_Cpp", configScope);
+
+    // Tag Parser if enabled else remove the setting with undefined
+    const value = isOptimizationEnabled? SETTING_TAG_PARSER: undefined;
+
+    await updateSetting(cppConfig, SETTING_INTELLISENSE_ENGINE, value);
+
+    if(isOptimizationEnabled){
+        console.warning("Optimization of UE is enabled!")
+        console.log("Enabling UE optimization disables red squiggle compiling for the Unreal Engine source code.");
+        console.log("This does not affect your project's source code (it still can have red squiggles).");
+        console.log("You can disable this optimization in this extension's settings.")
     }
-
-    renameUE4InConfig(ue4CCppPropertiesFirstConfig, ProjectUE4.ue4Version);
-
-    removeCompileCommandsFromConfig(allUE4CppPropertiesConfigs);
-
-    addBasicUE4IntellisenseToConfig(allUE4CppPropertiesConfigs);
-
-    await addResponsePreIncludesToMainWorkspaceConfig(project);
-
+    else {
+        console.log("Optimization of UE is disabled.")
+    }
+    
     return;
 }
 
 
-function renameUE4InConfig(outConfig: CCppConfigurationJson, ue4Version: string) {
+async function updateSetting(cppConfig: vscode.WorkspaceConfiguration, setting: string, value: string | undefined) {
 
-    outConfig.name = `Unreal Engine ${ue4Version} Source`;
-}
+    const currentValue = cppConfig.inspect(setting)?.workspaceFolderValue;
 
-
-function removeCompileCommandsFromConfig(outConfigs: CCppConfigurationJson[]) {
-    for (const config of outConfigs) {
-        config.compileCommands = undefined;
-    }
-    
-}
-
-
-function addBasicUE4IntellisenseToConfig(outConfigs: CCppConfigurationJson[]) {
-    for (const config of outConfigs) {
-        config.includePath = [consts.VSCODE_SPECIAL_VAR_DEFAULT].concat(consts.CONFIG_VALUES_GENERIC_UE4_INCLUDE_PATHS);
-    }
-    
-}
-
-
-async function addResponsePreIncludesToMainWorkspaceConfig(project: ProjectUE4) {
-
-    const preincludeMatches = await findPreincludeMatches(project);
-
-    if (!preincludeMatches?.length) {
-        console.error("Couldn't find preinclude matches in response file.");
+    if(currentValue === value){
         return;
     }
 
-    const extractedForcedIncludePaths = [consts.VSCODE_SPECIAL_VAR_DEFAULT].concat(preincludeMatches);
-
-    const mainSettings: WorkspaceConfiguration | undefined = project.getCCppSettingsConfig(project.mainWorkspaceKey);
-
-    if (!mainSettings) {
-        console.error("Couldn't get main settings. Can't change forced includes.");
+    try {
+        await cppConfig.update(SETTING_INTELLISENSE_ENGINE, value, null);
+    } catch (error) {
+        console.error(`Exception while updating ${SETTING_INTELLISENSE_ENGINE } setting with ${value}: UE Optimization`)
         return;
-    }
-
-    const currentForcedIncludes: string[] | undefined = mainSettings.get(consts.CONFIG_SETTING_DEFAULT_FORCED_INCLUDE);
-
-    checkAndUpdateForcedIncludes(mainSettings, currentForcedIncludes, extractedForcedIncludePaths);
-}
-
-
-async function findPreincludeMatches(project: ProjectUE4): Promise< RegExpMatchArray | undefined | null > {
-    const mainCompileCommands = project.getMainFirstConfigCompileCommands();
-
-    const responsePaths = mainCompileCommands?.getAllUsedResponsePaths();
-
-    if (!responsePaths) {
-        console.error(`Response paths was ${responsePaths}. Can't add preincludes.`);
-        return;
-    }
-
-    let matches : RegExpMatchArray | undefined | null = undefined;
-
-    for (const responsePath of responsePaths){
-        const responseFileString = await readStringFromFile(responsePath);
-        
-        if(!responseFileString){
-            continue;
-        }
-        
-        const match = responseFileString.match(consts.RE_COMPILE_COMMAND_FORCED_PATHS);
-        if(!match?.length){
-            continue;  // TODO warning message here???
-        }
-
-        if(!matches){ 
-            matches = match;
-        }
-        else{
-            matches = matches.concat(match);
-        }
-    }
-
-    return matches;
-    
-}
-
-
-function checkAndUpdateForcedIncludes( mainSettings: WorkspaceConfiguration, currentForcedIncludes: string[] | undefined, extractedForcedIncludePaths: string[]){
-    if (!currentForcedIncludes) {
-        console.log("Updating main workspace file's forced includes.");
-        mainSettings.update(consts.CONFIG_SETTING_DEFAULT_FORCED_INCLUDE, extractedForcedIncludePaths, false);
-        return;
-    }
-    else {
-        if (extractedForcedIncludePaths.every(value => {
-            return currentForcedIncludes.includes(value);
-        })) {
-            console.log("Current Main workspace file's forced includes are already fixed.");
-            return;
-        }
-        else {
-            console.log("Overwriting main workspace file's forced includes.");
-            mainSettings.update(consts.CONFIG_SETTING_DEFAULT_FORCED_INCLUDE, extractedForcedIncludePaths, false);
-            return;
-        }
     }
 }