+++ src/extension/fixes/propCompilerPath.ts
@@ -8,30 +8,43 @@ import { isMacM1 } from "../../shared";
 
 import * as console from "../../console";
 
-export function fixPropCompilerPath(project: ProjectUE4) {
+export async function fixPropCompilerPath(project: ProjectUE4) {
     
     console.log("Fixing compiler path in c_cpp_properties.json.");
 
     const config = vscode.workspace.getConfiguration(consts.CONFIG_SECTION_EXTENSION_COMPILER);
     let currentExtCompilerPath = config.get<string>(consts.CONFIG_SETTINGS_PATH);
 
-    if(!currentExtCompilerPath) {
 
-        if (isMacM1()){
-            console.log("Extension compiler path isn't set. Mac M1 compiler path will be auto set.");
-            config.update(consts.CONFIG_SETTINGS_PATH, consts.MACM1_DEFAULT_COMPILER_PATH, true);
-            currentExtCompilerPath = consts.MACM1_DEFAULT_COMPILER_PATH;
+    if(!currentExtCompilerPath && isMacM1()){
+        console.log("Extension compiler path isn't set. Mac M1 compiler path will be auto set.");
+
+        try {
+            await config.update(consts.CONFIG_SETTINGS_PATH, consts.MACM1_DEFAULT_COMPILER_PATH, true);
+        } catch (error) {
+            console.error("Error updating M1 Mac compiler path setting!")
         }
-        else{
-            console.log("Extension compiler path isn't set. Will not force compiler path in c_cpp_properties. (Not an error)");
-            console.log("If the extension's compiler path setting was previously set you should reset your UE project to get back the c_cpp_properties default path.\n");
-            return;
-        }
-        
+           
+        currentExtCompilerPath = consts.MACM1_DEFAULT_COMPILER_PATH;  // Even with error we can still set it below
     }
 
-    console.log(`Extension compiler path is set to: ${currentExtCompilerPath}`);
+    console.log(`Extension compiler path is set to: ${currentExtCompilerPath} (can be blank)`);
 
+    if(!currentExtCompilerPath) {
+
+            console.log("Extension compiler path isn't set. Will not force compiler path in c_cpp_properties. (Not an error)");
+            console.log("If the extension's compiler path setting was previously set you should reset your UE project to get back the c_cpp_properties default path.\n");      
+    }
+    else {
+
+        setCCppPropCompilerPathToExtPath(project, currentExtCompilerPath);
+    }
+
+    console.log("End fix c_cpp_properties compiler path.\n");
+}
+
+
+function setCCppPropCompilerPathToExtPath(project: ProjectUE4, currentExtCompilerPath: string) {
     const mainCppProp = project.getCCppConfigurationsFromWorkspace(project.mainWorkspaceKey);
     const ue4CppProp = project.getCCppConfigurationsFromWorkspace(project.ue4WorkspaceKey);
 
@@ -46,5 +59,4 @@ export function fixPropCompilerPath(project: ProjectUE4) {
         }        
     }
 
-    console.log("End fix c_cpp_properties compiler path.\n");
 }